analytics - Refactor DBConnection as QueryBuilder
DBConnection had three responsibilities:
1- buildings the final query out of a parametrized sql
  query and the parameter mappings;
2- executing the built query returning a Resultset;
3- and ensuring the sanity of the Connection object.

It's now refactored as QueryBuilder with the sole
responsibility of creating the final SQL query.

Point 2 above implied to return a ResultSet, and
the statement and the resultset were never closed.
It was now changed so that the client code asks
QueryBuilder to create the final query and is in
charge of executing it and disposing all resources.

Point 3 above was an anti-pattern, since keeping a single
connection that's going to be concurrently accessed will
result in concurrency issues (Connections are not thread safe).
Additionally, the responsibility of checking for the sanity
of the Connections is on the connection pool.